feature: modern bundle
