--- babel.config.js
@@ -1,13 +1,0 @@
-module.exports = {
-    presets: [
-        [
-            '@babel/preset-env',
-            {
-                targets: {
-                    node: 'current',
-                    edge: '18'
-                },
-            },
-        ],
-    ],
-};
